+++ lib/groupdate/relation.rb
@@ -11,7 +11,8 @@ module Groupdate
     def calculate(*args, &block)
       # TODO in next major version
       # pass default based on operation
-      Groupdate.process_result(self, super, default_value: 0)
+      default_value = [:count, :sum].include?(args[0]) ? 0 : nil
+      Groupdate.process_result(self, super, default_value: default_value)
     end
   end
 end